move cv to root component
