+++ src/app/work/page.tsx
@@ -1,16 +1,7 @@
 import Link from "next/link";
-import { WorkItem } from "./components/WorkItem";
-import { WORK_HISTORY_DATA } from "./constants";
 import { PageLayout } from "@/components/PageLayout";
 import { PageTitle } from "@/components/PageTitle";
-
-export interface WorkItemData {
-  company?: string;
-  jobTitle: string;
-  description?: string;
-  dateRange: string;
-  tools?: string[];
-}
+import { CurriculumVitae } from "@/components/CurriculumVitae";
 
 export default function WorkPage() {
   return (
@@ -25,19 +16,7 @@ export default function WorkPage() {
           </Link>
         </p>
       </div>
-      <div className="lg:overflow-scroll">
-        <ul className="list-none">
-          {WORK_HISTORY_DATA.map((item, index) => {
-            return (
-              <WorkItem
-                key={`${item.company}${item.jobTitle}`}
-                isLastIndex={index + 1 === WORK_HISTORY_DATA.length}
-                item={item}
-              />
-            );
-          })}
-        </ul>
-      </div>
+      <div className="lg:overflow-scroll">{<CurriculumVitae />}</div>
     </PageLayout>
   );
 }
